Remove the old module and move the test to the new module
